+++ CHANGELOG.md
@@ -22,6 +22,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
 - add option to not overwrite existing flags to `concatFlags`
 - add option to pass existing axis object to `plot`
 - python 3.11 support
+- added Local Outlier Factor functionality
 ### Changed
 - Remove all flag value restrictions from the default flagging scheme `FloatTranslator` 
 - Renamed `TranslationScheme.forward` to `TranslationScheme.toInternal`